Do not unpack jobs for identity processes
Also fixes waiting on lock file and cleans up using trap on EXIT.

+++ knit-dispatch-job.sh
@@ -15,33 +15,29 @@ job="$2"
 
 scratch="$KNIT_DIR/scratch/$job"
 
-exec 4> "$scratch.lock"
-if ! flock -n 4; then
-    echo "Waiting for: $(<$scratch.lock)" >&3
-fi
-if [[ -s "$scratch.lock" ]]; then
-    echo "warning: stale lockfile $scratch.lock: $(<$scratch.lock)" >&2
-    truncate -s 0 "$scratch.lock"
-fi
-# TODO handle manual processes
-echo "PID $$ session $session" >&4
+lock_scratch() {
+    if [[ -s "$scratch.lock" ]]; then
+        echo "warning: existing lockfile $scratch.lock: $(< $scratch.lock)" >&2
+    fi
+    exec 4> "$scratch.lock"
+    flock 4
+    trap 'rm "$scratch.lock"' EXIT
+    # TODO handle manual processes
+    echo "PID $$ session $session" >&4
+}
 
-if [[ -e $scratch ]]; then
-    echo "warning: removing $scratch" >&2
-    rm -rf "$scratch"
-fi
+unpack_job() {
+    if [[ -e $scratch ]]; then
+        echo "warning: removing $scratch" >&2
+        rm -rf "$scratch"
+    fi
 
-if prd=$(knit-cache "$job"); then
-    echo "Cache hit (locked) $job -> $prd" >&3
-    knit-complete-job "$session" "$job" "$prd"
-    rm "$scratch.lock"
-    exit
-fi
+    knit-unpack-job --scratch "$job" "$scratch"
+}
 
-# TODO remove job unpacking for identity process
-knit-unpack-job --scratch "$job" "$scratch"
+process_cmd() {
+    local rc
 
-if [[ -e $scratch/work/in/.knit/cmd ]]; then
     set +e
     # TODO disambiguate errors from knit-exec-cmd and .knit/cmd
     knit-exec-cmd "$scratch" "$scratch/work/in/.knit/cmd" 3>&- 4>&- > "$scratch/out.knit/log"
@@ -64,20 +60,40 @@ if [[ -e $scratch/work/in/.knit/cmd ]]; then
 
     rm -rf "$scratch/work/job"
     echo "$job" > "$scratch/work/job"
-    prd=$(knit-pack-production "$scratch/work")
-elif [[ -e $scratch/work/in/.knit/identity ]]; then
-    empty_res=$(knit-hash-object -t resource -w /dev/null)
-    prd=$(knit-remix-production --set-job "$job" --copy-job-inputs "$job" \
-        --remove-prefix .knit/ --set-output ".knit/ok=$empty_res")
-else
-    echo "Unsupported job $job" >&2
-    exit 1
-fi
+    knit-pack-production "$scratch/work"
+}
 
-echo "Complete $job -> $prd" >&3
-knit-cache "$job" "$prd"
-knit-complete-job "$session" "$job" "$prd"
+finish() {
+    echo "Complete $job -> $prd" >&3
+    knit-cache "$job" "$prd"
+    knit-complete-job "$session" "$job" "$prd"
+    exit
+}
 
-# TODO when to keep scratch dir?
-rm -rf "$scratch"
-rm "$scratch.lock"
+unset prd
+while read -r input; do
+    if [[ $input == .knit/cmd ]]; then
+        lock_scratch
+        if prd=$(knit-cache "$job"); then
+            echo "Cache hit (locked) $job -> $prd" >&3
+            knit-complete-job "$session" "$job" "$prd"
+            rm "$scratch.lock"
+            exit
+        fi
+        unpack_job
+
+        prd=$(process_cmd)
+
+        # TODO when to keep scratch dir?
+        rm -rf "$scratch"
+        finish
+    elif [[ $input == .knit/identity ]]; then
+        empty_res=$(knit-hash-object -t resource -w /dev/null)
+        prd=$(knit-remix-production --set-job "$job" --copy-job-inputs "$job" \
+            --remove-prefix .knit/ --set-output ".knit/ok=$empty_res")
+        finish
+    fi
+done < <(knit-cat-file -p "$job" | cut -f2-)
+
+echo "Unsupported job $job" >&2
+exit 1
